+++ common/src/main/java/dev/ithundxr/createnumismatics/registry/commands/DeductCommand.java
@@ -1,6 +1,7 @@
 package dev.ithundxr.createnumismatics.registry.commands;
 
 import com.mojang.authlib.GameProfile;
+import com.mojang.brigadier.Command;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
 import com.simibubi.create.foundation.blockEntity.behaviour.BlockEntityBehaviour;
@@ -42,6 +43,12 @@ public class DeductCommand {
                             BlockPos pos = BlockPosArgument.getLoadedBlockPos(ctx, "pos");
                             int amount = getInteger(ctx, "amount");
                             BankAccountBehaviour bankAct = BlockEntityBehaviour.get(ctx.getSource().getLevel(), pos, BankAccountBehaviour.TYPE);
+
+                            if (bankAct == null) {
+                                ctx.getSource().sendFailure(Components.literal("There is no Blaze Banker at " + pos.toShortString()));
+                                return -1;
+                            }
+
                             UUID id = bankAct.getAccountUUID();
 
                             return execute(ctx, id, Type.BLAZE_BANKER, false, "Blaze Banker at (" + pos.toShortString() + ")", amount, force);
@@ -52,6 +59,12 @@ public class DeductCommand {
                                 int amount = getInteger(ctx, "amount");
                                 Coin coin = ctx.getArgument("coin", Coin.class);
                                 BankAccountBehaviour bankAct = BlockEntityBehaviour.get(ctx.getSource().getLevel(), pos, BankAccountBehaviour.TYPE);
+
+                                if (bankAct == null) {
+                                    ctx.getSource().sendFailure(Components.literal("There is no Blaze Banker at " + pos.toShortString()));
+                                    return -1;
+                                }
+
                                 UUID id = bankAct.getAccountUUID();
 
                                 return execute(ctx, id, Type.BLAZE_BANKER, false, "Blaze Banker at (" + pos.toShortString() + ")", amount, force, coin);