+++ common/src/main/java/dev/ithundxr/createnumismatics/registry/commands/ViewCommand.java
@@ -32,6 +32,12 @@ public class ViewCommand {
                     .executes(ctx -> {
                         BlockPos pos = BlockPosArgument.getLoadedBlockPos(ctx, "pos");
                         BankAccountBehaviour bankAct = BlockEntityBehaviour.get(ctx.getSource().getLevel(), pos, BankAccountBehaviour.TYPE);
+
+                        if (bankAct == null) {
+                            ctx.getSource().sendFailure(Components.literal("There is no Blaze Banker at " + pos.toShortString()));
+                            return -1;
+                        }
+
                         UUID id = bankAct.getAccountUUID();
 
                         return execute(ctx, id, Type.BLAZE_BANKER, false, "Blaze Banker at (" + pos.toShortString() + ")");
@@ -41,6 +47,12 @@ public class ViewCommand {
                             BlockPos pos = BlockPosArgument.getLoadedBlockPos(ctx, "pos");
                             Coin coin = ctx.getArgument("coin", Coin.class);
                             BankAccountBehaviour bankAct = BlockEntityBehaviour.get(ctx.getSource().getLevel(), pos, BankAccountBehaviour.TYPE);
+
+                            if (bankAct == null) {
+                                ctx.getSource().sendFailure(Components.literal("There is no Blaze Banker at " + pos.toShortString()));
+                                return -1;
+                            }
+
                             UUID id = bankAct.getAccountUUID();
 
                             return execute(ctx, id, Type.BLAZE_BANKER, false, "Blaze Banker at (" + pos.toShortString() + ")", coin);